fix(corfu): remove conflict with `+org/return`
Previously, when pressing Enter in org-mode, a newline would always be
inserted, regardless of corfu state. This led to annoying behavior: you
could select a candidate and preview it in the buffer via corfu, but RET
would erase its overlay and break the line instead of inserting the
candidate.

An alternative way to fix this would be to advise `+org/return` in
"modules/completion/corfu/config.el", which would make this commit more
self-contained, but I find that to be too hacky.